Move files from legal-data-preprocessing-cfr

+++ download_us_reg_data.py
@@ -1,27 +1,31 @@
+import os
 import shutil
 from multiprocessing.pool import Pool
-from zipfile import ZipFile
 
 import requests
 
 from statics import US_REG_INPUT_PATH
-from utils.common import ensure_exists
+
+
+def ensure_exists(path):
+    if not os.path.exists(path):
+        os.makedirs(path)
+    return path
+
 
 DOWNLOAD_BASE_URL = "https://www.govinfo.gov/bulkdata/CFR/{}/CFR-{}.zip"
 
 
 def download(year):
-    print("loading", year)
-    r = requests.get(DOWNLOAD_BASE_URL.format(year, year), stream=True)
-    if r.status_code == 200:
-        zip_path = f"{US_REG_INPUT_PATH}/{year}.zip"
-        with open(zip_path, "wb") as f:
-            r.raw.decode_content = True
-            shutil.copyfileobj(r.raw, f)
-        print("downloaded", year)
-
-        with ZipFile(zip_path) as f:
-            f.extractall(ensure_exists(f"{US_REG_INPUT_PATH}/{year}"))
+    zip_path = f"{US_REG_INPUT_PATH}/{year}.zip"
+    if not os.path.exists(zip_path):
+        print("loading", year)
+        r = requests.get(DOWNLOAD_BASE_URL.format(year, year), stream=True)
+        if r.status_code == 200:
+            with open(zip_path, "wb") as f:
+                r.raw.decode_content = True
+                shutil.copyfileobj(r.raw, f)
+            print("downloaded", year)
 
 
 if __name__ == "__main__":
